factor out main ch-frb-simulate-l0 functionality
